% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{as.cell_data_set}
\alias{as.cell_data_set}
\alias{as.CellDataSet}
\alias{as.cell_data_set.Seurat}
\title{Convert objects to Monocle3 \code{cell_data_set} objects}
\usage{
as.cell_data_set(x, ...)

\method{as.cell_data_set}{Seurat}(
  x,
  assay = DefaultAssay(object = x),
  reductions = AssociatedDimReducs(object = x, assay = assay),
  default.reduction = DefaultDimReduc(object = x, assay = assay),
  graph = paste0(assay, "_snn"),
  group.by = NULL,
  ...
)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{Assay to convert}

\item{reductions}{A vector of dimensional reductions add to the
\code{cell_data_set} object; defaults to all dimensional reductions
calculated from \code{assay} and all \link[Seurat:IsGlobal]{global}
dimensional reductions}

\item{default.reduction}{Name of dimensional reduction to use for clustering
name}

\item{graph}{Name of graph to be used for clustering results}

\item{group.by}{Name of cell-level metadata column to use as identites; pass}
}
\value{
A \code{cell_data_set} object
}
\description{
Convert objects to Monocle3 \code{cell_data_set} objects
}
