% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{as.Seurat.list}
\alias{as.Seurat.list}
\alias{as.Seurat}
\title{Create a Seurat object from a list of matricies}
\usage{
\method{as.Seurat}{list}(x, default.assay = 1, slot = "counts",
  min.cells = 0, min.features = 0, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object to convert to class \code{Seurat}}

\item{default.assay}{Name or index of matrix to use as default assay;
defaults to name of first matrix in list}

\item{slot}{Name of slot to store matrix in; choose from 'counts' or 'data'}

\item{min.cells}{Include features detected in at least this many cells. Will
subset the counts matrix as well. To reintroduce excluded features, create a
new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{verbose}{Display progress updates}

\item{...}{Arguments passed to other methods}
}
\description{
Create a Seurat object from a list of matricies
}
\seealso{
\code{\link[Seurat]{as.Seurat}}
}
