% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralysis.R
\name{CoralysisIntegration}
\alias{CoralysisIntegration}
\title{Coralysis integration in Seurat v5}
\usage{
CoralysisIntegration(
  object,
  batch,
  threads = 1,
  ndims = 50,
  assay = NULL,
  layers = NULL,
  features = NULL,
  new.reduction = "integrated.dr",
  reduction.key = "coralysis_",
  reconstructed.assay = "coralysis",
  save.coralysis.metadata = NULL,
  verbose = TRUE,
  verbose.icp = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object with joint layers.}

\item{batch}{Cell metadata column name representing the batch label identity 
to guide integration. By default uses the default assay.}

\item{threads}{Threads to pass on to \code{Coralysis}. By default it is not 
performed parallelization.}

\item{ndims}{A positive integer denoting the number of principal components 
to calculate and select. Default is \code{50}.}

\item{assay}{Seurat assay to retrieve layers from. By default uses the default 
assay.}

\item{layers}{Seurat layers to retrieve the normalized expression matrix from. 
By default uses the layer \code{data}, i.e., the normalized expression data.}

\item{features}{A vector of features to use for integration. Otherwise 2,000
highly variable features are internally selected by default. Set the argument 
of \code{scale.layer} to \code{NULL} in the function \code{IntegrateLayers}  
to use a set of \code{features} other than default.}

\item{new.reduction}{Name of the new integrated dimensional reduction.}

\item{reduction.key}{Reduction key for the new (integrated) reduction.}

\item{reconstructed.assay}{Name of the integrated assay comprising the cell 
cluster probabilities from Coralysis.}

\item{save.coralysis.metadata}{Filename to save Coralysis metadata locally 
as an RDS object. Useful to re-use the ICP models generated during integration. 
By default it is not saved.}

\item{verbose}{Logical value indicating whether to print progress messages.}

\item{verbose.icp}{Logical value indicating whether to print progress messages 
during the ICP modeling step. Defaults to \code{FALSE}, meaning no messages 
are printed. Useful to debug ICP.}

\item{...}{Arguments passed on to \code{Coralysis}. Check possible arguments 
with \code{?Coralysis::RunParallelDivisiveICP}.}
}
\value{
A list with the new integrated reduction (\code{new.reduction}) and 
the integrated assay comprising the cell cluster probabilities from Coralysis
(\code{reconstructed.assay}).
}
\description{
Run Coralysis integration in Seurat v5.
}
\note{
This function requires the \href{https://github.com/elolab/Coralysis}{\pkg{Coralysis}} 
package to be installed.
}
\examples{
\dontrun{
# Basic Seurat workflow
seu <- SeuratData::LoadData("ifnb")
seu[["RNA"]] <- split(seu[["RNA"]], f = seu$stim)
seu <- FindVariableFeatures(seu)
seu <- ScaleData(seu)
seu <- RunPCA(seu)

# Coralysis integration with default parameters 
seu <- IntegrateLayers(object = seu, 
  method = CoralysisIntegration,
  new.reduction = "integrated.coralysis",
  batch = "stim",
  threads = 4)
  
 # Coralysis integration with other than default parameters 
 # (see other options `?Coralysis::RunParallelDivisiveICP`)
seu <- IntegrateLayers(object = seu, 
  method = CoralysisIntegration,
  new.reduction = "integrated.coralysis",
  batch = "stim",
  threads = 4, 
  k = 8, 
  L = 25)
}

}
\references{
Sousa A, Smolander J, Junttila S, Elo L (2025). “Coralysis enables 
sensitive identification of imbalanced cell types and states in single-cell data 
via multi-level integration.” bioRxiv. doi:10.1101/2025.02.07.637023
}
\seealso{
\code{\link[Coralysis]{RunParallelDivisiveICP}} \code{\link[Coralysis]{RunPCA}}
}
\author{
António GG Sousa
}
