% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbrowser.R
\name{findMatrices}
\alias{findMatrices}
\title{used by ExportToCellbrowser:
Return a matrix object from a Seurat object or show an error message}
\usage{
findMatrices(object, slotNames)
}
\arguments{
\item{object}{Seurat object}

\item{slotNames}{vector with the names of the slots to export}
}
\value{
a list with filename prefix -> matrix object. prefix usually is "scale_data" or "counts"
}
\description{
used by ExportToCellbrowser:
Return a matrix object from a Seurat object or show an error message
}
