% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbrowser.R
\name{findMatrix}
\alias{findMatrix}
\title{used by ExportToCellbrowser:
Return a matrix object from a Seurat object or show an error message}
\usage{
findMatrix(object, matrix.slot)
}
\arguments{
\item{object}{Seurat object}

\item{matrix.slot}{the name of the slot}
}
\description{
used by ExportToCellbrowser:
Return a matrix object from a Seurat object or show an error message
}
