% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttopics.R
\name{FitPoissonNMF}
\alias{FitPoissonNMF}
\title{Add Title Here.}
\usage{
FitPoissonNMF(
  object,
  k,
  assay = NULL,
  features = NULL,
  reduction.name = "poisson_nmf",
  reduction.key = "k_",
  numiter = 100,
  method = c("scd", "em", "mu", "ccd"),
  init.method = c("topicscore", "random"),
  control = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{k}{An integer 2 or greater giving the matrix rank. This
argument should only be specified if the initial fit (\code{fit0})
is not provided.}

\item{assay}{Name of assay to use; defaults to the default
assay of the object.}

\item{reduction.name}{Name of the outputted reduction.}

\item{reduction.key}{Key for the outputted reduction.}

\item{numiter}{}

\item{verbose}{When \code{verbose = TRUE}, information about the
progress of the model fitting is printed to the console. See
\code{\link[fastTopics]{fit_poisson_nmf}} for an explanation of the
output.}

\item{\dots}{Additional arguments passed to \code{fit_poisson_nmf};
see \code{\link[fastTopics]{fit_poisson_nmf}} for details.}
}
\value{
Describe the return value here.
}
\description{
Add description here.
}
\details{
Add details here.
}
\examples{
library(Seurat)
library(fastTopics)
set.seed(1)
data(pbmc_small)

# Fit the non-negative matrix factorization to the raw UMI count
# data; no pre-processing or pre-selection of genes is needed.
pbmc_small <- FitPoissonNMF(pbmc_small,k = 3)

}
\references{
Lee, D. D. and Seung, H. S. (2001). Algorithms for non-negative
  matrix factorization. In \emph{Advances in Neural Information
  Processing Systems} \bold{13}, 556–562.
}
\seealso{
\code{\link{FitTopicModel}},
  \code{\link[fastTopics]{fit_poisson_nmf}}
}
\author{
Peter Carbonetto
}
