% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttopics.R
\name{FitTopicModel}
\alias{FitTopicModel}
\title{Add Title Here}
\usage{
FitTopicModel(
  object,
  k = 3,
  assay = NULL,
  features = NULL,
  reduction.name = "fasttopics",
  reduction.key = "k_",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Describe input argument "object" here.}

\item{k}{Describe input argument "k" here.}

\item{assay}{Describe input argument "assay" here.}

\item{features}{Describe input argument "features" here.}

\item{reduction.name}{Describe input argument "reduction.name" here.}

\item{reduction.key}{Describe input argument "reduction.key" here.}

\item{verbose}{Describe input argument "verbose" here.}

\item{\dots}{Describe "..." here.}
}
\value{
Describe the return value here.

Clarify that, unfortunately, "factors" and "loadings" in fastTopics
mean the opposite of what they mean in Seurat.
}
\description{
Add description here.
}
\seealso{
\code{\link[fastTopics]{fit_topic_model}}
}
