% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{LearnGraph}
\alias{LearnGraph}
\title{Run \code{link[monocle3]{learn_graph}} on a \code{\link[Seurat]{Seurat}} object}
\usage{
LearnGraph(object, reduction = DefaultDimReduc(object = object), ...)
}
\arguments{
\item{object}{A \code{\link[Seurat]{Seurat}} object}

\item{reduction}{Name of reduction to use for learning the pseudotime graph}

\item{...}{Arguments passed to \code{\link[monocle3]{learn_graph}}}
}
\value{
A \code{\link[monocle3]{cell_data_set}} object with the pseudotime graph
}
\description{
Run \code{link[monocle3]{learn_graph}} on a \code{\link[Seurat]{Seurat}} object
}
\seealso{
\code{\link[monocle3]{learn_graph}} \code{\link[monocle3]{cell_data_set}}
}
