% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexbin.R
\name{make_hexbin}
\alias{make_hexbin}
\title{Bivariate binning of single cell data into hexagon cells.}
\usage{
make_hexbin(sce, nbins = 80, dimension_reduction = "UMAP")
}
\arguments{
\item{sce}{A \code{\link[Seurat]{Seurat}} object.}

\item{nbins}{The number of bins partitioning the range of the first
component of the chosen dimension reduction.}

\item{dimension_reduction}{A string indicating the reduced dimension
result to calculate hexagon cell representation of.}
}
\value{
A \code{\link[Seurat]{Seurat}} object.
}
\description{
\code{make_hexbin} returns a
  \code{\link[Seurat]{Seurat}} object of binned hexagon cells.
}
\details{
This function bins observations with computed reduced dimension
  results into hexagon cells. For a \code{\link[Seurat]{Seurat}} object the
  results from this function are stored in \code{slot(object, "misc")}.
  The list contains two items. The first item stores a vector specifying the
  hexagon ID for each observation. The second item stores a matrix with the
  x and y positions of the hexagon cells and the number of observations in
  each of them.
}
\references{
Delile, Julien, et al.
   "Single cell transcriptomics reveals spatial and temporal dynamics of gene
    expression in the developing mouse spinal cord." Development 146.12
    (2019): dev173807.

 @examples
# For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
}
\seealso{
\code{\link[schex]{make_hexbin}}
}
\author{
Saskia Freytag
}
