% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexbin_label.R
\name{make_hexbin_label}
\alias{make_hexbin_label}
\title{Group label plot position.}
\usage{
make_hexbin_label(sce, col)
}
\arguments{
\item{sce}{A \code{\link[Seurat]{Seurat}} object.}

\item{col}{The name referring to one column in meta data for which
the label position on the plot is calculated for every level. The chosen
column needs to be a factor.}
}
\value{
A tibble.
}
\description{
Group label plot position.
}
\examples{
#' # For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
make_hexbin_label(pbmc_small, col="RNA_snn_res.1")
}
\references{
Delile, Julien, et al.
   "Single cell transcriptomics reveals spatial and temporal dynamics of gene
    expression in the developing mouse spinal cord." Development 146.12
    (2019): dev173807.
}
\seealso{
\code{\link[schex]{make_hexbin_label}}
}
\author{
Saskia Freytag
}
