% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttopics.R
\name{PerformGoMDEAnalysis}
\alias{PerformGoMDEAnalysis}
\title{Perform a Differential Expression Analysis using a Topic Model}
\usage{
PerformGoMDEAnalysis(object, ...)
}
\arguments{
\item{object}{A Seurat object containing a previously fitted
multinomial topic model or Poisson NMF dimensionality reduction
object.}

\item{\dots}{Additional arguments passed to
\code{\link[fastTopics]{de_analysis}}.}
}
\value{
Describe the return value here.
}
\description{
Add description here.
}
\seealso{
\code{\link{FitPoissonNMF}},
\code{\link{FitTopicModel}}
}
