% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_density.R
\name{plot_hexbin_density}
\alias{plot_hexbin_density}
\title{Plot of density of observations from single cell data in bivariate hexagon
  cells.}
\usage{
plot_hexbin_density(sce, title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[Seurat]{Seurat}} object.}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of density of observations from single cell data in bivariate hexagon
  cells.
}
\examples{
#' # For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
plot_hexbin_density(pbmc_small)
}
\references{
Delile, Julien, et al.
   "Single cell transcriptomics reveals spatial and temporal dynamics of gene
    expression in the developing mouse spinal cord." Development 146.12
    (2019): dev173807.
}
\seealso{
\code{\link[schex]{plot_hexbin_density}}
}
\author{
Saskia Freytag
}
