% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_gene.R
\name{plot_hexbin_gene}
\alias{plot_hexbin_gene}
\title{Plot of gene expression of single cells in bivariate hexagon cells.}
\usage{
plot_hexbin_gene(sce, type, gene, action, title = NULL, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[Seurat]{Seurat}} object.}

\item{type}{A string referring to the type of expression data plotted.
Possible options are \code{counts} for raw counts and \code{logcounts}
for the scaled data in the \code{\link[Seurat]{Seurat}} object.}

\item{gene}{A string referring to the name of one gene.}

\item{action}{A strings pecifying how meta data of observations in
binned  hexagon cells are to be summarized. Possible actions are
\code{prop_0}, \code{mode}, \code{mean} and \code{median} (see details).}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of gene expression of single cells in bivariate hexagon cells.
}
\details{
This function plots the expression of any gene in the hexagon cell
  representation calculated with \code{\link{make_hexbin}}. The chosen gene
  expression is summarized by one of four actions \code{prop_0}, \code{mode},
  \code{mean} and \code{median}:

  \describe{
    \item{\code{prop_0}}{Returns the proportion of observations in the bin
     greater than 0. The associated meta data column needs to be numeric.}
    \item{\code{mode}}{Returns the mode of the observations in the bin. The
     associated meta data column needs to be numeric.}
    \item{\code{mean}}{Returns the mean of the observations in the bin. The
     associated meta data column needs to be numeric.}
     \item{\code{median}}{Returns the median of the observations in the bin.
     The associated meta data column needs to be numeric.}
  }
}
\examples{
# For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
plot_hexbin_gene(pbmc_small, type="counts", gene="TALDO1", action="prop_0")
}
\references{
Delile, Julien, et al.
   "Single cell transcriptomics reveals spatial and temporal dynamics of gene
    expression in the developing mouse spinal cord." Development 146.12
    (2019): dev173807.
}
\seealso{
\code{\link[schex]{plot_hexbin_gene}}
}
\author{
Saskia Freytag
}
