% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_meta.R
\name{plot_hexbin_meta}
\alias{plot_hexbin_meta}
\title{Plot of meta data of single cell data in bivariate hexagon cells.}
\usage{
plot_hexbin_meta(sce, col, action, no = 1, colors = NULL,
  title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[Seurat]{Seurat}} object.}

\item{col}{A string referring to the name of one column in the meta data of
sce by which to colour the hexagons.}

\item{action}{A string specifying how meta data of observations in
binned  hexagon cells are to be summarized. Possible actions are
\code{majority}, \code{prop}, \code{prop_0}, \code{mode}, \code{mean} and
\code{median} (see details).}

\item{no}{An integer specifying which level to plot of the column. Only in
effect when \code{action=prop}.}

\item{colors}{A vector of strings specifying which colors to use for plotting
the different levels in the selected column of the meta data. Only in
effect when the selected \code{action="majority"}.}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of meta data of single cell data in bivariate hexagon cells.
}
\details{
This function plots any column of the meta data in the hexagon cell
  representation calculated with \code{\link{make_hexbin}}. The chosen meta
  data column is summarized by one of six actions \code{majority},
  \code{prop}, \code{prop_0}, \code{mode}, \code{mean} and \code{median}:

  \describe{
    \item{\code{majority}}{Returns the value of the majority of observations
     in the bin. The associated meta data column needs to be a factor
     or character.}
    \item{\code{prop}}{Returns the proportion of each level or unique
     character in the bin. The associated meta data column needs to be a
     factor or character.}
    \item{\code{prop_0}}{Returns the proportion of observations in the bin
     greater than 0. The associated meta data column needs to be numeric.}
    \item{\code{mode}}{Returns the mode of the observations in the bin. The
     associated meta data column needs to be numeric.}
    \item{\code{mean}}{Returns the mean of the observations in the bin. The
     associated meta data column needs to be numeric.}
     \item{\code{median}}{Returns the median of the observations in the bin.
     The associated meta data column needs to be numeric.}
  }
}
\examples{
#' # For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
plot_hexbin_meta(pbmc_small, col="RNA_snn_res.1", action="prop", no=1)
plot_hexbin_meta(pbmc_small, col="RNA_snn_res.1", action="prop", no=2)
}
\references{
Delile, Julien, et al.
   "Single cell transcriptomics reveals spatial and temporal dynamics of gene
    expression in the developing mouse spinal cord." Development 146.12
    (2019): dev173807.
}
\seealso{
\code{\link[schex]{plot_hexbin_meta}}
}
\author{
Saskia Freytag
}
