% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqc.R
\name{PlotMiQC}
\alias{PlotMiQC}
\title{Run miQC on a Seurat object}
\usage{
PlotMiQC(
  seurat_object,
  percent.mt = "percent.mt",
  nFeature_RNA = "nFeature_RNA",
  model.slot = "flexmix_model",
  color.by = "miQC.probability"
)
}
\arguments{
\item{object}{Seurat object}
}
\value{

}
\description{
Run miQC on a Seurat object
}
\details{
_Function to plot the miQC mixture model stored in a Seurat object. `RunMiQC` must be run prior to plotting._
}
\references{
Hippen et al. (2021) miQC: An adaptive probabilistic framework for quality control of single-cell RNA-sequencing data. bioRxiv doi: 10.1101/2021.03.03.433798
}
