% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevin.R
\name{ReadAlevin}
\alias{ReadAlevin}
\title{Load alevin quantification data}
\usage{
ReadAlevin(file, ...)
}
\arguments{
\item{file}{path to \code{quants_mat.gz} file within
alevin directory}

\item{...}{extra arguments passed to \code{tximport},
for example,
\code{alevinArgs=list(filterBarcodes=TRUE, tierImport=TRUE)}.}
}
\value{
returns a Seurat object with alevin counts
}
\description{
A wrapper around tximport to create a \code{SeuratObject}
from alevin quantification data.
}
\references{
Srivastava, Avi, et al. "Alevin efficiently 
estimates accurate gene abundances from dscRNA-seq data." 
Genome biology 20.1 (2019): 65.
}
\seealso{
\code{\link[alevin]{alevin}}
}
\author{
Avi Srivastava
}
