% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{ReadVelocity}
\alias{ReadVelocity}
\title{Load RNA Velocity data from a loom file}
\usage{
ReadVelocity(file, engine = "hdf5r", verbose = TRUE)
}
\arguments{
\item{file}{Path to loom file}

\item{engine}{Method to load data data, choose from 'hdf5r' or 'h5'}

\item{verbose}{Display progress updates}
}
\description{
This is a wrapper around \code{\link[velocyto.R]{read.loom.matrices}}, but sends
messages to \code{stderr} instead of \code{stdout} (or silences messages with
\code{verbose = FALSE})
}
\seealso{
\code{\link[velocyto.R]{read.loom.matrices}}
}
