% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogaps.R
\name{RunCoGAPS}
\alias{RunCoGAPS}
\title{Run CoGAPs on a Seurat object}
\usage{
RunCoGAPS(
  object,
  assay = NULL,
  slot = "counts",
  params = NULL,
  temp.file = NULL,
  reduction.name = "CoGAPS",
  reduction.key = "CoGAPS_",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay to pull data from}

\item{slot}{Slot to pull data from.}

\item{params}{\code{\link[CoGAPS]{CogapsParams}} object for specifying parameter settings}

\item{temp.file}{Name of temporary data matrix file to create if running in a distributed mode.
Setting to TRUE will generate the file name using \code{tempfile}.}

\item{reduction.name}{Name of the CoGAPS reduction returned}

\item{reduction.key}{Key for the CoGAPS reduction returned}
}
\value{
Returns a Seurat object with the CoGAPS results stored as a \code{\link{DimReduc}} object
}
\description{
Run CoGAPs on a Seurat object
}
\references{
E.J. Fertig, J. Ding, A.V. Favorov, G. Parmigiani, and M.F. Ochs (2010) CoGAPS: an
integrated R/C++ package to identify overlapping patterns of activation of biological processes
from expression data. Bioinformatics 26:2792-2793.
}
\seealso{
\code{\link[CoGAPS]{CoGAPS}}
}
