% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{RunMAGIC}
\alias{RunMAGIC}
\title{Run Magic}
\usage{
RunMAGIC(
  object,
  assay = "RNA",
  new.assay.name = "MAGIC",
  genes = "all_genes",
  knn = 15,
  solver = c("exact", "approximate"),
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay used as input to MAGIC, default RNA}

\item{new.assay.name}{New created assay from MAGIC output}

\item{genes}{Genes imputed for, all genes by default}

\item{knn}{knn number for MAGIC algorithm}

\item{solver}{solver used in MAGIC}
}
\value{
Seurat object with new assay named MAGIC from MAGIC output
}
\description{
Run Magic
}
\seealso{
\link[Rmagic package]{https://cran.r-project.org/web/packages/Rmagic/index.html}
}
