% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqc.R
\name{RunMiQC}
\alias{RunMiQC}
\title{Run miQC on a Seurat object}
\usage{
RunMiQC(
  object,
  percent.mt = "percent.mt",
  nFeature_RNA = "nFeature_RNA",
  posterior.cutoff = 0.75,
  model.type = "linear",
  model.slot = "flexmix_model",
  verbose = TRUE,
  backup.option = "percentile",
  backup.percentile = 0.99,
  backup.percent = 5,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{percent.mt}{(character) Name of the column in the Seurat metadata that
contains the percent of reads attributed to mitochondrial genes.
Defaults to "percent.mt".}

\item{nFeature_RNA}{(character) Name of the column in the Seurat metadata that
contains the number of reads per cell. Defaults to "nFeature_RNA".}

\item{posterior.cutoff}{numeric) The posterior probability of a cell being
part of the compromised distribution, a number between 0 and 1. Any cells
below the appointed cutoff will be marked to keep.
Defaults to 0.75.}

\item{model.type}{(character) What type of model to generate. A linear
mixture model ("linear") is recommended, but currently b-spline ("spline")
and two-degree polynomial ("polynomial") are also supported
Default = "linear".}

\item{verbose}{Boolean. TRUE to show progress messages, FALSE to hide progress messages}

\item{backup.option}{(character) In case flexmix fails to build a 2 cluster
mixture model, what should RunMiQC do: "percent" (set miQC.keep values
according to backup.percent), "percentile" (set miQC.keep values according
to backup.percentile), "pass" (return original Seurat object), or "halt"
(stop RunMiQC). "percent", "percentile", and "pass" are useful when
processing multiple Seurat objects sequentially.}

\item{backup.percentile}{(numeric) What percentile to use as cutoff in case
flexmix fails to build a 2 cluster mixture model. Will only be used if
backup.option is "percentile".}

\item{backup.percent}{(numeric) What percent to use as cutoff in case flexmix
fails to build a 2 cluster mixture model. Will only be used if
backup.option is "percent".}
}
\value{
Returns a Seurat object with probabilities and "keep" decisions stored as "miQC.probability" and "miQC.keep" in the object metadata, respectively.
}
\description{
Run miQC on a Seurat object
}
\details{
(Copied verbatim from miQC) _Function to fit a two-distribution mixture model on a Seurat object and find those cells probabistically determined to be compromised by the mixture model._
}
\references{
Hippen et al. (2021) miQC: An adaptive probabilistic framework for quality control of single-cell RNA-sequencing data. bioRxiv doi: 10.1101/2021.03.03.433798
}
