% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto.R
\name{RunPresto}
\alias{RunPresto}
\title{A Presto-based implementation of FindMarkers that runs Wilcoxon tests for the given identity classes}
\usage{
RunPresto(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  subset.ident = NULL,
  assay = NULL,
  slot = "data",
  reduction = NULL,
  features = NULL,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  min.pct = 0.1,
  min.diff.pct = -Inf,
  verbose = TRUE,
  only.pos = FALSE,
  max.cells.per.ident = Inf,
  random.seed = 1,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  pseudocount.use = 1,
  mean.fxn = NULL,
  fc.name = NULL,
  base = 2,
  ...
)
}
\arguments{
\item{ident.1}{Identity class to define markers for; pass an object of class
\code{phylo} or 'clustertree' to find markers for a node in a cluster tree;
passing 'clustertree' requires \code{\link{BuildClusterTree}} to have been run}

\item{ident.2}{A second identity class for comparison; if \code{NULL},
use all other cells for comparison; if an object of class \code{phylo} or
'clustertree' is passed to \code{ident.1}, must pass a node to find markers for}

\item{group.by}{Regroup cells into a different identity class prior to performing differential expression (see example)}

\item{subset.ident}{Subset a particular identity class prior to regrouping. Only relevant if group.by is set (see example)}

\item{assay}{Assay to use in differential expression testing}

\item{slot}{Slot to pull data from; note that if \code{test.use} is "negbinom", "poisson", or "DESeq2",
\code{slot} will be set to "counts"}

\item{reduction}{Reduction to use in differential expression testing - will test for DE on cell embeddings}

\item{mean.fxn}{Function to use for fold change or average difference calculation.
If NULL, the appropriate function will be chose according to the slot used}

\item{fc.name}{Name of the fold change, average difference, or custom function column
in the output data.frame. If NULL, the fold change column will be named
according to the logarithm base (eg, "avg_log2FC"), or if using the scale.data
slot "avg_diff".}

\item{base}{The base with respect to which logarithms are computed.}
}
\description{
A Presto-based implementation of FindMarkers that runs Wilcoxon tests for the given identity classes
}
\seealso{
https://github.com/immunogenomics/presto
}
