% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto.R
\name{RunPrestoAll}
\alias{RunPrestoAll}
\alias{RunPrestoAllNode}
\title{A Presto-based implementation of FindAllMarkers that runs Wilcoxon tests for all identity classes}
\usage{
RunPrestoAll(
  object,
  assay = NULL,
  features = NULL,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  slot = "data",
  min.pct = 0.1,
  min.diff.pct = -Inf,
  node = NULL,
  verbose = TRUE,
  only.pos = FALSE,
  max.cells.per.ident = Inf,
  random.seed = 1,
  latent.vars = NULL,
  min.cells.feature = 3,
  min.cells.group = 3,
  pseudocount.use = 1,
  mean.fxn = NULL,
  fc.name = NULL,
  base = 2,
  return.thresh = 0.01,
  ...
)
}
\arguments{
\item{assay}{Assay to use in differential expression testing}

\item{slot}{Slot to pull data from; note that if \code{test.use} is "negbinom", "poisson", or "DESeq2",
\code{slot} will be set to "counts"}

\item{node}{A node to find markers for and all its children; requires
\code{\link{BuildClusterTree}} to have been run previously; replaces \code{FindAllMarkersNode}}

\item{mean.fxn}{Function to use for fold change or average difference calculation.
If NULL, the appropriate function will be chose according to the slot used}

\item{fc.name}{Name of the fold change, average difference, or custom function column
in the output data.frame. If NULL, the fold change column will be named
according to the logarithm base (eg, "avg_log2FC"), or if using the scale.data
slot "avg_diff".}

\item{base}{The base with respect to which logarithms are computed.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)}
}
\value{
Matrix containing a ranked list of putative markers, and associated
statistics (p-values, logFC, etc.)
}
\description{
Finds markers (Wilcoxon-differentially expressed genes) for each of the identity classes in a dataset
}
\seealso{
https://github.com/immunogenomics/presto
}
