% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGSEA.R
\name{RunScGSEA}
\alias{RunScGSEA}
\title{Run Single cell Gene Set Enrichement Analysis on GF-ICF on a Seurat object}
\usage{
RunScGSEA(
  object,
  assay = NULL,
  filterGenes = TRUE,
  nmf.dim = 100,
  geneID = c("ensamble", "symbol"),
  species = c("mouse", "human"),
  category = "H",
  subcategory = NULL,
  verbose = F
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay to use, defaults to the default assay}

\item{filterGenes}{Rank of NMF to use for the enrichment}

\item{nmf.dim}{Rank of NMF to use for the enrichment}

\item{geneID}{The type of gene names in rownames of \code{object}. It can be either 'ensamble' or 'symbol'. Default: 'ensamble'}

\item{species}{The type of species in \code{object}. It can be either 'mouse' or 'human'. Default: 'human'}

\item{category}{MSigDB collection abbreviation, such as H or C1}

\item{subcategory}{MSigDB sub-collection abbreviation, such as CGP or BP}

\item{rescale}{If different by none, pathway's activity scores are resealed as Z-score. Possible values are none, byGS or byCell. Default is none}
}
\value{
returns a Seurat object with pathways x cell matrix
}
\description{
This function run runScGSEA function from GF-ICF package on Seurat object. 
It computes GSEA for each cells across a set of input pathways by using NMF.
}
