% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricycle.R
\name{Runtricycle}
\alias{Runtricycle}
\title{Run estimate_cycle_position on a Seurat object}
\usage{
Runtricycle(
  object,
  assay = NULL,
  slot = "data",
  reduction.name = "tricycleEmbedding",
  reduction.key = "tricycleEmbedding_",
  gname = NULL,
  gname.type = c("ENSEMBL", "SYMBOL"),
  species = c("mouse", "human"),
  AnnotationDb = NULL,
  center.pc1 = 0,
  center.pc2 = 0
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay to use, defaults to the default assay}

\item{slot}{Slot to use. It should be library size adjusted **log-expression** values.
Note that it is convention that we rename "logcounts" to "data" when converting SingleCellExperiment to Seurat object.
See also \code{\link[Seurat]{as.Seurat}}. Defaults to "data"}

\item{reduction.name}{Name of the cell cycle projection returned}

\item{reduction.key}{Key for the cell cycle projection returned}

\item{gname}{Alternative rownames of \code{object}. If provided, this will be used to map genes within \code{object} with genes in reference.
If not provided, the rownames of \code{object} will be used instead. Default: NULL}

\item{gname.type}{The type of gene names as in \code{gname} or rownames of \code{object}. It can be either 'ENSEMBL' or 'SYMBOL'. Default: 'ENSEMBL'}

\item{species}{The type of species in \code{object}. It can be either 'mouse' or 'human'. Default: 'mouse'}

\item{AnnotationDb}{An AnnotationDb objects. If the user provides rownames in the format of Ensembl IDs and project human data,
this object will be used to map Ensembl IDs to gene SYMBOLs. If no AnnotationDb object being given, the function will use \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}.}

\item{center.pc1}{The center of PC1 when defining the angle. Default: 0}

\item{center.pc2}{The center of PC2 when defining the angle. Default: 0}
}
\description{
This function run estimate_cycle_position function on Seurat object. It uses
the tricycle internal reference projection matrix.
}
