% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{RunVelocity}
\alias{RunVelocity}
\title{Run RNA Velocty}
\usage{
RunVelocity(
  object,
  spliced = "spliced",
  unspliced = "unspliced",
  ambiguous = NULL,
  spliced.average = 0.2,
  unspliced.average = 0.05,
  reduction = "pca",
  group.by = "ident",
  cells = NULL,
  graph = NULL,
  ncores = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{spliced}{Name of spliced assay}

\item{unspliced}{Name of unspliced assay}

\item{ambiguous}{Optional name of ambiguous assay}

\item{spliced.average, unspliced.average}{Required minimum average expression count for the spliced and unspliced expression matrices}

\item{reduction}{Name of reduction to use}

\item{group.by}{Factor to group cells by}

\item{cells}{Vector of cells to use; defaults to all cells
(see \code{\link[velocyto.R]{gene.relative.velocity.estimates}:steady.state.cells})}

\item{graph}{Optional name of nearest neighbor graph to use}

\item{ncores}{Number of cores to use}

\item{verbose}{Display progress updates}

\item{...}{Extra parameters passed to \code{\link[velocyto.R]{gene.relative.velocity.estimates}}}
}
\value{
...
}
\description{
Run RNA Velocty
}
\seealso{
\code{\link[velocyto.R]{gene.relative.velocity.estimates}} \code{\link[Seurat]{Tool}}
}
