% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbrowser.R
\name{saveMatrix}
\alias{saveMatrix}
\title{used by ExportToCellbrowser:
Write a matrix to a file, as either as .tsv.gz or .mtx.gz+barcodes+genes
Filename is exprMatrix<suffix>.tsv.gz}
\usage{
saveMatrix(counts, dir, prefix, use.mtx)
}
\arguments{
\item{counts}{The matrix, usually a sparse matrix}

\item{prefix}{Added to the base filename, can be "" or "norm_" or something similar}

\item{use.mtx}{If true, will write to <prefix>matrix.mtx and <prefix>features/barcodes.tsv}
}
\description{
used by ExportToCellbrowser:
Write a matrix to a file, as either as .tsv.gz or .mtx.gz+barcodes+genes
Filename is exprMatrix<suffix>.tsv.gz
}
