% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanorama.R
\name{ScanoramaIntegration}
\alias{ScanoramaIntegration}
\title{Run scanorama in Seurat5}
\usage{
ScanoramaIntegration(
  object,
  orig,
  layers = NULL,
  features = NULL,
  scale.layer = "scale.data",
  conda_env = NULL,
  new.reduction = "integrated.scanorama",
  reduction.key = "scanorama_",
  reconstructed.assay = "scanorama.reconstructed",
  ndims.out = 100L,
  return_dense = TRUE,
  batch_size = 5000,
  approx = TRUE,
  sigma = 15L,
  alpha = 0.1,
  knn = 20L,
  hvg.scanorama = NULL,
  union.features = FALSE,
  sketch = FALSE,
  sketch_method = c("geosketch", "uniform"),
  sketch_max = 10000,
  seed.use = 42L,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{\link[SeuratObject]{Assay5}} object}

\item{orig}{A \link[SeuratObject:DimReduc]{dimensional reduction} to correct}

\item{layers}{Layers to look for. Either a vector of layer names or a single
string (e.g. 'counts' or 'data')}

\item{features}{A vector of features to use for integration.
All features used by default}

\item{conda_env}{Path to conda environment to run \pkg{scanorama} (should also
contain the scipy python module)}

\item{new.reduction}{Name to store resulting
\link[SeuratObject:DimReduc]{dimensional reduction} object. \code{NULL} or
\code{FALSE} disables the dimension reduction computation}

\item{reduction.key}{Key for resulting
\link[SeuratObject:DimReduc]{dimensional reduction} object. Ignored if
\code{reduction.key} is \code{NULL} or \code{FALSE}}

\item{reconstructed.assay}{Name for the \code{assay} containing the corrected
expression matrix of dimension #\code{features} x #cells. \code{NULL} or
\code{FALSE} disables the corrected expression matrix computation}

\item{ndims.out}{Number of dimensions for \code{new.reduction} output.
Scanorama specific argument}

\item{return_dense}{Whether \pkg{scanorama} returns \code{numpy.ndarray} matrices
instead of \code{scipy.sparse.csr_matrix}. Scanorama specific argument}

\item{batch_size}{Used in the alignment vector computation. Higer values
increase memory burden.Scanorama specific argument}

\item{approx}{Whether scanorama approximates nearest neighbors computation.
Scanorama specific argument}

\item{sigma}{Correction smoothing parameter on gaussian kernel. Scanorama
specific argument}

\item{alpha}{Minimum alignment score. Scanorama specific argument}

\item{knn}{Number of nearest neighbours used for matching. Scanorama
specific argument}

\item{hvg.scanorama}{A positive to turn \pkg{scanorama}'s internal HVG selection on.
Disabled by default. Scanorama specific argument}

\item{union.features}{By default, \pkg{scanorama} uses the intersection of features
to perform integration. Set this parameter to \code{TRUE} to use the union.
Discouraged. Scanorama specific argument}

\item{sketch}{Turns on sketching-based acceleration by first downsampling the
datasets. See Hie et al., Cell Systems (2019). Disabled by default. Ignored
when \code{reconstructed.assay} is enabled. Scanorama specific argument}

\item{sketch_method}{A skething method to apply to the data. Either
'geosketch' (default) or 'uniform'. Ignored when \code{reconstructed.assay}
is enabled or when \code{sketch} is FALSE. Scanorama specific argument}

\item{sketch_max}{Downsampling cutoff. Ignored when sketching disabled.
Scanorama specific argument}

\item{seed.use}{Value of \pkg{scanorama}'s internal seed. Set to \code{NULL} to
disable}

\item{verbose}{Print messages}

\item{...}{Ignored}
}
\value{
A list containing at least one of:
\itemize{
  \item a new DimReduc of name \code{reduction.name} (key set to
  \code{reduction.key}) with corrected embeddings matrix of \code{ndims.out}.
  \item a new Assay of name \code{reconstructed.assay} with corrected counts
  for each \code{features} (or less if \code{hvg} is set to a lower integer).
}
}
\description{
Integrate batches using \pkg{scanorama} in \link[Seurat:Assay5]{Seurat's Assay5} through \link[Seurat]{IntegrateLayers}
}
\note{
This function requires the
\href{https://github.com/brianhie/scanorama}{\pkg{Scanorama}} package
to be installed (along with \pkg{scipy}) with conda
}
\examples{
\dontrun{
# Preprocessing
obj <- UpdateSeuratObject(SeuratData::LoadData("pbmcsca"))
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

# After preprocessing, we integrate layers:
obj <- IntegrateLayers(object = obj, method = ScanoramaIntegration,
                       # change conda_env value to your conda env with scanorama
                       conda_env = '~/miniconda3/envs/scanorama')
print(Reductions(obj))
print(Assays(obj))

# To disable feature expression matrix correction:
obj <- IntegrateLayers(object = obj, method = ScanoramaIntegration,
                       # change conda_env value to your conda env with scanorama
                       conda_env = '~/miniconda3/envs/scanorama',
                       reconstructed.assay = NULL)

# To disable reduced dimension correction:
obj <- IntegrateLayers(object = obj, method = ScanoramaIntegration,
                       # change conda_env value to your conda env with scanorama
                       conda_env = '~/miniconda3/envs/scanorama',
                       new.reduction = NULL)
}

}
\references{
Hie, B., Bryson, B. & Berger, B. Efficient integration of
heterogeneous single-cell transcriptomes using Scanorama. Nat Biotechnol 37,
685-691 (2019). \href{https://doi.org/10.1038/s41587-019-0113-3}{DOI}
}
\seealso{
\code{\link[Seurat]{Tool}}
}
