% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbrowser.R
\name{writeSparseTsvChunks}
\alias{writeSparseTsvChunks}
\title{Used by \code{ExportToCellbrowser}:
Write a big sparse matrix to a .tsv.gz file by writing chunks, concating them with the Unix cat command,
then gziping the result. This does not work on Windows, we'd have to use the copy /b command there.}
\usage{
writeSparseTsvChunks(inMat, outFname, sliceSize = 1000)
}
\arguments{
\item{inMat}{input matrix}

\item{outFname}{output file name, has to end with .gz}

\item{sliceSize=1000, }{size of each chunk in number of lines}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Used by \code{ExportToCellbrowser}:
Write a big sparse matrix to a .tsv.gz file by writing chunks, concating them with the Unix cat command,
then gziping the result. This does not work on Windows, we'd have to use the copy /b command there.
}
\examples{
\dontrun{
writeSparseTsvChunks( pbmc_small@data, "exprMatrix.tsv.gz")
}

}
