% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbrowser.R
\name{CellBrowser}
\alias{CellBrowser}
\alias{ExportToCellbrowser}
\title{Export \code{Seurat} objects for UCSC cell browser and stop open cell browser
instances from R}
\usage{
ExportToCellbrowser(
  object,
  dir,
  dataset.name = Project(object = object),
  reductions = NULL,
  markers.file = NULL,
  cluster.field = NULL,
  cb.dir = NULL,
  port = NULL,
  use.mtx = FALSE,
  meta.fields = NULL,
  meta.fields.names = NULL,
  matrix.slot = "counts",
  markers.n = 100,
  skip.markers = FALSE,
  skip.expr.matrix = FALSE,
  skip.metadata = FALSE,
  skip.reductions = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{dir}{path to directory where to save exported files. These are:
exprMatrix.tsv, tsne.coords.tsv, meta.tsv, markers.tsv and a default
cellbrowser.conf}

\item{dataset.name}{name of the dataset. Defaults to Seurat project name}

\item{reductions}{vector of reduction names to export, defaults to all reductions.}

\item{markers.file}{path to file with marker genes. By defaults, marker
are searched in the object itself as misc$markers. If none are supplied in
object or via this argument, they are recalculated with \code{FindAllMarkers}}

\item{cluster.field}{name of the metadata field containing cell cluster}

\item{cb.dir}{path to directory where to create UCSC cellbrowser static
website content root, e.g. an index.html, .json files, etc. These files
can be copied to any webserver. If this is specified, the cellbrowser
package has to be accessible from R via reticulate.}

\item{port}{on which port to run UCSC cellbrowser webserver after export}

\item{use.mtx}{export the matrix in .mtx.gz format. Default is False,
unless the matrix is bigger than R's maximum matrix size.}

\item{meta.fields}{vector of meta fields to export, default is all.}

\item{meta.fields.names}{vector meta field names to show in UI. Must have
same length as meta.fields. Default is meta.fields.}

\item{matrix.slot}{matrix to use, default is 'counts'}

\item{markers.n}{if no markers were supplied, FindAllMarkers is run.
This parameter indicates how many markers to calculate, default is 100}

\item{skip.markers}{whether to skip exporting markers}

\item{skip.expr.matrix}{whether to skip exporting expression matrix}

\item{skip.metadata}{whether to skip exporting metadata}

\item{skip.reductions}{whether to skip exporting reductions}

\item{...}{specifies the metadata fields to export. To supply a field and its
human readable name, pass name as \code{field="name"} parameter.}
}
\value{
This function exports Seurat object as a set of tsv files
to \code{dir} directory, copying the \code{markers.file} if it is
passed. It also creates the default \code{cellbrowser.conf} in the
directory. This directory could be read by \code{cbBuild} to
create a static website viewer for the dataset. If \code{cb.dir}
parameter is passed, the function runs \code{cbBuild} (if it is
installed) to create this static website in \code{cb.dir} directory.
If \code{port} parameter is passed, it also runs the webserver for
that directory and opens a browser.
}
\description{
Export \code{Seurat} objects for UCSC cell browser and stop open cell browser
instances from R
}
\examples{
\dontrun{
ExportToCellbrowser(pbmc_small, dataset.name = "PBMC", dir = "out")
}

}
\author{
Maximilian Haeussler, Nikolay Markov
}
