% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttopics.R
\name{FitTopicModel}
\alias{FitTopicModel}
\title{Fit a Multinomial Topic Model using fastTopics}
\usage{
FitTopicModel(
  object,
  k = 3,
  assay = NULL,
  features = NULL,
  reduction.name = "multinom_topic_model",
  reduction.key = "k_",
  reduction.pca.name = "pca_topics",
  reduction.pca.key = "TOPICPC_",
  verbose = c("progressbar", "detailed", "none"),
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{k}{The number of topics. Must be 2 or more.}

\item{assay}{Name of assay to use; defaults to the default
assay of the object.}

\item{features}{A list of features to use for fitting the model. If
\code{features = NULL}, \emph{all} features are used; in
particular, \code{\link[Seurat]{VariableFeatures}} is \emph{not}
used to pre-select features.}

\item{reduction.name}{Name of the outputted reduction.}

\item{reduction.key}{Key for the outputted reduction.}

\item{reduction.pca.name}{Name of the outputted PCA reduction.}

\item{reduction.pca.key}{Key for the outputted PCA reduction.}

\item{verbose}{When \code{verbose = "progressbar"} or \code{verbose
= "detailed"}, information about the progress of the model fitting
is printed to the console. See
\code{\link[fastTopics]{fit_poisson_nmf}} for more information.}

\item{\dots}{Additional arguments passed to \code{fit_topic_model};
see \code{\link[fastTopics]{fit_topic_model}} for details.}
}
\value{
A Seurat object, in which the multinomial topic model fit
is stored as a Seurat \code{\link[Seurat]{DimReduc}} object. The
cell embeddings (stored in the \code{cell.embeddings} slot) are the
topic proportions; this is the n x k matrix \code{L} outputted by
\code{fit_topic_model} (n is the number of cells, k is the
number of topics).

The feature loadings (stored in the \code{feature.loadings} slot)
are an m x k matrix of relative expression levels, where m is the
number of features and k is the number of topics; this is the
matrix \code{F} outputted by \code{fit_topic_model}.

Apply \code{\link[Seurat]{Misc}} to the \code{DimReduce} object to
access the \dQuote{multinom_topic_model_fit} object outputted by
\code{\link[fastTopics]{fit_topic_model}} which contains more
information about the topic model fit; see the example for an
illustration, and see \code{\link[fastTopics]{fit_topic_model}} more
information about the \dQuote{multinom_topic_model_fit} object.

An additional PCA dimension reduction computed from the topic
proportions is provided. This could be useful, for example, to
quickly visualize the cells using \code{\link[Seurat]{DimPlot}}. The
principal components are computed from the topic mixture
proportions. However, this reduction is provided for convenience
only, and we recommend to extract the
\dQuote{multinom_topic_model_fit} object and use the dedicated
visualization tools that are provided in fastTopics such as
\code{\link[fastTopics]{structure_plot}}.
}
\description{
Fits a multinomial topic model to the raw count data.
  The default optimization settings used here are intended to work
  well in a wide range of data sets, although some fine-tuning may be
  needed for larger or more complex data sets.
}
\details{
The topic model can be viewed as a dimensionality reduction
in which the topic proportions matrix, \code{L} (an n x k matrix,
where n is the number of cells and k is the number of topics),
defines a projection of the cells onto a k-dimensional embedding
(strictly speaking, it is a (k-1)-dimensional embedding).  The topic
proportions define the \dQuote{cell embedding} in the
\code{DimReduc} object.

\emph{A warning about terminology:} In fastTopics, the \code{L}
matrix is sometimes called the \dQuote{loadings matrix}. However,
this is not the same as the \dQuote{feature loadings} matrix in
Seurat. To avoid confusion, we instead prefer to refer to \code{L}
as the topic proportions matrix.

See \code{\link[fastTopics]{fit_topic_model}} for more details.
}
\examples{
library(Seurat)
library(fastTopics)
set.seed(1)
data(pbmc_small)

# Fit the multinomial topic model to the raw UMI count data; no
# pre-processing or pre-selection of genes is needed.
pbmc_small <- FitTopicModel(pbmc_small,k = 3)

# This plot shows the cells projected onto the 2 principal
# components (PCs) of the topic mixture proportions.
DimPlot(pbmc_small,reduction = "pca_topics")

# Compare this against the top two PCs of the transformed count
# data.
DimPlot(pbmc_small,reduction = "pca")

# Once fitted topic model is extracted, many functions from the
# fastTopics package can be used for analysis and visualization. For
# example, the Structure plot provides an evocative visual summary of
# the estimated mixture proportions for each cell.
fit <- Misc(Reductions(pbmc_small,"multinom_topic_model"))
structure_plot(fit,grouping = Idents(pbmc_small),gap = 5)

}
\references{
K. K. Dey, C. J. Hsiao and M. Stephens (2017). Visualizing the
structure of RNA-seq expression data using grade of membership
models. \emph{PLoS Genetics} \bold{13}, e1006599.

P. Carbonetto, A. Sarkar, Z. Wang and M. Stephens (2021).
Non-negative matrix factorization algorithms greatly improve
topic model fits. \emph{arXiv} 2105.13440.
}
\seealso{
\code{\link{FitPoissonNMF}},
  \code{\link{PerformGoMDEAnalysis}},
  \code{\link[fastTopics]{fit_topic_model}}
}
\author{
Peter Carbonetto
}
