% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttopics.R
\name{FitTopicModel}
\alias{FitTopicModel}
\title{Fit a Multinomial Topic Model Using fastTopics}
\usage{
FitTopicModel(
  object,
  k = 3,
  assay = NULL,
  features = NULL,
  reduction.name = "multinom_topic_model",
  reduction.key = "k_",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{k}{The number of topics. Must be 2 or more.}

\item{assay}{Name of assay to use; defaults to the default
assay of the object.}

\item{reduction.name}{Name of the outputted reduction.}

\item{reduction.key}{Key for the outputted reduction.}

\item{verbose}{When \code{verbose = TRUE}, information about the
progress of the model fitting is printed to the console. See
\code{\link[fastTopics]{fit_poisson_nmf}} for an explanation of the
output.}

\item{\dots}{Additional arguments passed to \code{fit_topic_model};
see \code{\link[fastTopics]{fit_topic_model}} for details.}
}
\value{
A Seurat object, in which the multinomial topic model fit
is stored as a Seurat \code{\link[Seurat]{DimReduc}} object. The
cell embeddings (stored in the \code{cell.embeddings} slot) are the
mixture proportions; this is the n x k matrix \code{L} outputted by
\code{fit_topic_model} (n is the number of cells and k is the
number of topics).

The feature loadings (stored in the \code{feature.loadings} slot)
are an m x k matrix of relative expression levels; this is the
matrix \code{F} outputted by \code{fit_topic_model}.

Apply \code{\link[Seurat]{Misc}} to the \code{DimReduce} object to
access the \dQuote{"multinom_topic_model_fit"} object outputted by
\code{\link[fastTopics]{fit_topic_model}}, which contains more
information about the multinomial topic model fit; see the example
here for an illustration of how to access the model fit, and see
\code{\link[fastTopics]{fit_topic_model}} more information about
the "multinom_topic_model_fit" object.

An additional PCA dimension reduction on \code{k-1} dimensions is
provided. This could be useful, for example, to quickly visualize
the cells using \code{\link[Seurat]{DimPlot}}. The principal
components are computed from the topic mixture
proportions. However, this reduction is provided for convenience
only, and we recommend to extract the
\dQuote{"multinom_topic_model_fit"} object and use the dedicated
visualization tools that are provided in the fastTopics package
such as \code{\link[fastTopics]{structure_plot}}.
}
\description{
Fits a multinomial topic model to the raw count data,
  hiding most of the complexities of model fitting. The default
  optimization settings used here are intended to work well in a wide
  range of data sets, although some fine-tuning may be needed for
  larger or more complex data sets.
}
\details{
The topic model can be understood as a dimensionality
reduction, in which the mixture proportions matrix, \code{L}, an n
x k matrix, defines a projection of the cells onto a
(k–1)-dimension space (n is the number of cells and k is the number
of topics). The topic mixture proportions define the \dQuote{cell
embedding} in the \code{DimReduc} object.

\emph{A warning about confusing terminology:} In fastTopics,
the \code{L} matrix is sometimes referred to as the "loadings
matrix" (because this is the convention used in factor
analysis). However, this is \emph{not} the same as the "feature
loadings" matrix used in Seurat, which adopts the convention used n
principal component analysis. To avoid confusion, we refer to
\code{L} as the mixture proportions matrix.

See \code{\link[fastTopics]{fit_topic_model}} for more details.
}
\examples{
library(Seurat)
library(fastTopics)
set.seed(1)

# Load the PBMC data.
data(pbmc_small)

# Fit the multinomial topic model to the raw UMI count data; no
# pre-processing or pre-selection of genes is needed.
pbmc_small <- FitTopicModel(pbmc_small,k = 3)

# This plot shows the cells projected onto the 2 principal
# components (PCs) of the topic mixture proportions.
DimPlot(pbmc_small,reduction = "pca_topics")

# Compare this against the top two PCs of the transformed count
# data.
DimPlot(pbmc_small,reduction = "pca")

# Once fitted topic model is extracted, many functions from the
# fastTopics package can be used for analysis and visualization. For
# example, the Structure plot provides an evocative visual summary of
# the estimated mixture proportions for each cell.
fit <- Misc(Reductions(pbmc_small,"multinom_topic_model"))
structure_plot(fit,grouping = Idents(pbmc_small),gap = 5)

}
\references{
Dey, K. K., Hsiao, C. J. and Stephens, M. (2017). Visualizing the
structure of RNA-seq expression data using grade of membership
models. \emph{PLoS Genetics} \bold{13}, e1006599.
}
\seealso{
\code{\link[fastTopics]{fit_topic_model}}
}
\author{
Peter Carbonetto
}
