% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alevin.R
\name{ReadAlevin}
\alias{ReadAlevin}
\title{Load alevin Counts data from an EDS file}
\usage{
ReadAlevin(file)
}
\arguments{
\item{file}{Path to `quants_mat.gz` file}
}
\value{
Returns a Seurat object with alevin counts
}
\description{
This is a wrapper around ReadCounts from AlevinRtools
}
\references{
Srivastava, Avi, et al. "Alevin efficiently 
estimates accurate gene abundances from dscRNA-seq data." 
Genome biology 20.1 (2019): 65.
}
\seealso{
\code{\link[alevin]{alevin}}
}
\author{
Avi Srivastava
}
