% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banksy.R
\name{RunBanksy}
\alias{RunBanksy}
\title{Run Banksy on a Seurat Object}
\usage{
RunBanksy(
  object,
  lambda,
  assay = "RNA",
  slot = "data",
  use_agf = FALSE,
  dimx = NULL,
  dimy = NULL,
  dimz = NULL,
  ndim = 2,
  features = "variable",
  group = NULL,
  split.scale = TRUE,
  k_geom = 15,
  n = 2,
  sigma = 1.5,
  alpha = 0.05,
  k_spatial = 10,
  spatial_mode = "kNN_median",
  assay_name = "BANKSY",
  M = NULL,
  chunk_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{lambda}{(numeric) Spatial weight parameter}

\item{assay}{(character) Assay in Seurat object to use}

\item{slot}{(character) Slot in Seurat assay to use}

\item{use_agf}{(boolean) Whether to use the AGF}

\item{dimx}{(character) Column name of spatial x dimension (must be in metadata)}

\item{dimy}{(character) Column name of spatial y dimension (must be in metadata)}

\item{dimz}{(character) Column name of spatial z dimension (must be in metadata)}

\item{ndim}{(integer) Number of spatial dimensions to extract}

\item{features}{(character) Features to compute. Can be 'all', 'variable' or
a vector of feature names}

\item{group}{(character) Column name of a grouping variable (must be in metadata)}

\item{split.scale}{(boolean) Whether to separate scaling by group}

\item{k_geom}{(numeric) kNN parameter - number of neighbors to use}

\item{n}{(numeric) kNN_rn parameter - exponent of radius}

\item{sigma}{(numeric) rNN parameter - standard deviation of Gaussian kernel}

\item{alpha}{(numeric) rNN parameter - determines radius used}

\item{k_spatial}{(numeric) rNN parameter - number of neighbors to use}

\item{spatial_mode}{(character) Kernel for neighborhood computation
\itemize{
 \item{kNN_median: k-nearest neighbors with median-scaled Gaussian kernel}
 \item{kNN_r: k-nearest neighbors with $1/r$ kernel}
 \item{kNN_rn: k-nearest neighbors with $1/r^n$ kernel}
 \item{kNN_rank: k-nearest neighbors with rank Gaussian kernel}
 \item{kNN_unif: k-nearest neighbors wth uniform kernel}
 \item{rNN_gauss: radial nearest neighbors with Gaussian kernel}
}}

\item{assay_name}{(character) Name for Banksy assay in Seurat object}

\item{M}{(numeric) Advanced usage. Highest azimuthal harmonic}

\item{chunk_size}{A integer scalar specifying the number of rows / genes of
the neighborhood cell matrix to compute. Must be less than floor of
2e31-1 / number of cells. This is automatically computed but can be
specified.#'}

\item{verbose}{(boolean) Print messages}
}
\value{
A Seurat object with new assay holding a Banksy matrix
}
\description{
Run Banksy on a Seurat Object
}
\references{
Vipul Singhal, Nigel Chou et. al. BANKSY: A Spatial Omics
Algorithm that Unifies Cell Type Clustering and Tissue Domain Segmentation
}
\seealso{
\code{\link[Banksy]{ComputeBanksy}}
}
\author{
Joseph Lee, Vipul Singhal
}
