% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{RunQuantileAlignSNF}
\alias{RunQuantileAlignSNF}
\alias{quantileAlignSNF}
\title{Run quantileAlignSNF on a Seurat object}
\usage{
RunQuantileAlignSNF(
  object,
  split.by = "orig.ident",
  reduction = "iNMF_raw",
  reduction.name = "iNMF",
  reduction.key = "iNMF_",
  recalc.snf = FALSE,
  ref_dataset = NULL,
  prune.thresh = 0.2,
  min_cells = 2,
  quantiles = 50,
  nstart = 10,
  resolution = 1,
  center = FALSE,
  id.number = NULL,
  print.mod = FALSE,
  print.align.summary = FALSE,
  ...
)
}
\arguments{
\item{object}{A merged Seurat object}

\item{split.by}{Attribute for splitting, defaults to "orig.ident"}

\item{reduction}{Name of reduction to use}

\item{reduction.name}{Name to store resulting DimReduc object as}

\item{reduction.key}{Key for resulting DimReduc}

\item{recalc.snf}{Recalculate \code{\link{SNF}}}

\item{...}{Arguments passed to other methods, and to
\code{\link[seurat.wrappers]{SNF}} if \code{recalc.snf = TRUE} or
\code{\link[seurat.wrappers]{SNF}} hasn't been run}
}
\value{
A Seurat object with embeddings from \code{\link[liger]{quantileAlignSNF}}
stored as a DimReduc object with name \code{reduction.name} (key set to \code{reduction.key})
}
\description{
This is a deprecated function. Call 'RunQuantileNorm' instead.
}
\seealso{
\code{\link[rliger]{RunQuantileNorm}}
}
