% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{RunQuantileNorm}
\alias{RunQuantileNorm}
\alias{quantile_norm}
\title{Run quantile_norm on a Seurat object}
\usage{
RunQuantileNorm(
  object,
  split.by = "orig.ident",
  reduction = "iNMF_raw",
  reduction.name = "iNMF",
  reduction.key = "iNMF_",
  quantiles = 50,
  ref_dataset = NULL,
  min_cells = 20,
  knn_k = 20,
  dims.use = NULL,
  do.center = FALSE,
  max_sample = 1000,
  eps = 0.9,
  refine.knn = TRUE,
  ...
)
}
\arguments{
\item{object}{A merged Seurat object}

\item{split.by}{Attribute for splitting, defaults to "orig.ident"}

\item{reduction.name}{Name to store resulting DimReduc object as}

\item{reduction.key}{Key for resulting DimReduc}

\item{...}{Arguments passed to other methods}
}
\value{
A Seurat object with embeddings from \code{\link[liger]{quantile_norm}}
stored as a DimReduc object with name \code{reduction.name} (key set to \code{reduction.key})
}
\description{
Run quantile_norm on a Seurat object
}
\seealso{
\code{\link[rliger]{quantile_norm}}
}
