% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVI.R
\name{scVIIntegration}
\alias{scVIIntegration}
\title{Run scVI in seurat5}
\usage{
scVIIntegration(
  object,
  groups = NULL,
  features = NULL,
  layers = "counts",
  conda_env = NULL,
  new.reduction = "integrated.dr",
  ndims = 10,
  max_epochs = 200,
  ...
)
}
\arguments{
\item{object}{A merged Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
A Seurat object with embeddings and loadings
}
\description{
Run scVI in seurat5
}
